remove /site from .gitignore
This was causing issues with using a directory called site on this branch. .gitignore does not "git add -A" anything listed in it, or in other words, ignores it. Usually .gitignore is used to ignore unnecessary files such as temporary files, binary files, build files, editor files (like VS code), or operating system code files.